toolchain: Enable deterministic archive
Used by ar, ranlib, objcopy, strip when operating on .a file. When
enabled, all timestamps, GID, UID is setted to zero for checksum
determinism.

It's not strictly required for reproducable binary but allows to
have deterministic .a files - it should be the same on every machine
when compiled from same source code.

JIRA: ISC-222