aarch64: add -mno-outline-atomics compilation flag
GCC 14.2 may implement atomic operations using a function that may use
LSE extension instructions. It decided whether to use these instructions
using a flag in memory. However, because we do not use glibc, the flag
was never initialized. The "outline atomics" feature would require
additional changes in kernel, libphoenix and plo and possibly toolchain.

Disabling the feature seems like a better choice because Phoenix-RTOS
applications are compiled for a specific platform, so the target
instruction set (and extensions) are known at compile time.

With -mno-outline-atomics flag, GCC only uses LSE instructions if they
are mandatory for the architecture given in -mcpu flag. LSE is mandatory
starting from Arm v8.1. If Phoenix-RTOS is ever ported to an Arm v8.0
system with LSE and enabling support would be beneficial, we can do it
per-target.

JIRA: RTOS-1050