scripts/image_builder: early return from parsing PLO script if 'if' clause is false
The PLO script was parsed first and then checked for 'if' clauses, which did not
give the script's user the possibility of e.g. checking if variable is defined
and using it in the command if that is the case. Checking before rendering gives
the user more flexibility.

Example: imxrt117x `user.plo.yaml` script could wants to check if env.LWIP_EXEC_CONFIG
is defined and use it in the command if that's the case.

JIRA: RTOS-963